+++ pypdfbox/pdmodel/fdf/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from .fdf_annotation import FDFAnnotation
 from .fdf_annotation_circle import FDFAnnotationCircle
+from .fdf_annotation_file_attachment import FDFAnnotationFileAttachment
 from .fdf_annotation_free_text import FDFAnnotationFreeText
 from .fdf_annotation_line import FDFAnnotationLine
 from .fdf_annotation_square import FDFAnnotationSquare
@@ -14,6 +15,7 @@ from .fdf_field import FDFField
 __all__ = [
     "FDFAnnotation",
     "FDFAnnotationCircle",
+    "FDFAnnotationFileAttachment",
     "FDFAnnotationFreeText",
     "FDFAnnotationLine",
     "FDFAnnotationSquare",